python-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-openai
openai
chromadb
sentence-transformers
pypdf
python-docx
tiktoken

langchain_huggingface

langchain_google_genai
sentence-transformers